refactor: Reorganize layout of LXD configuration classes
