--- src/cleantest/control/_configurers/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2023 Jason C. Nucciarone, Canonical Ltd.
-# See LICENSE file for licensing details.
-
-from .lxd_configurer import LXDConfigurer
