+++ src/cleantest/control/lxd/__init__.py
@@ -1,5 +1,7 @@
 #!/usr/bin/env python3
-# Copyright 2023 Jason C. Nucciarone, Canonical Ltd.
+# Copyright 2023 Jason C. Nucciarone
 # See LICENSE file for licensing details.
+
+"""Classes for configuring the LXD test environment provider."""
 
 from .lxd_config import ClientConfig, InstanceConfig, InstanceSource
