+++ src/cleantest/meta/mixins/__init__.py
@@ -1,6 +1,9 @@
 #!/usr/bin/env python3
-# Copyright 2023 Jason C. Nucciarone, Canonical Ltd.
+# Copyright 2023 Jason C. Nucciarone
 # See LICENSE file for licensing details.
 
+"""Public mixins used inside cleantest."""
+
 from .dict_ops import DictOps
+from .enhanced_enum import EnhancedEnum
 from .snapd_support import SnapdSupport
